# openapi.yaml
openapi: 3.0.0
info:
  title: fitlife API
  description: API REST completa con autenticación JWT y CRUD de entidades
  version: 1.0.0
servers:
  - url: http://localhost:3003/api
    description: Servidor de desarrollo
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    EntityBase:
      type: object
      properties:
        _id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        imagen:
          type: string
        imagenThumbnail:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
...
paths:
...
  /plan-nutricional:
    get:
      tags:
        - PlanNutricional
      summary: Listar todos los PlanNutricionals
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de PlanNutricionals
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - PlanNutricional
      summary: Crear PlanNutricional
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: PlanNutricional creado
  /plan-nutricional/{id}:
    get:
      tags:
        - PlanNutricional
      summary: Obtener PlanNutricional por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PlanNutricional encontrado
        '404':
          description: PlanNutricional no encontrado
    patch:
      tags:
        - PlanNutricional
      summary: Actualizar PlanNutricional
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: PlanNutricional actualizado
    delete:
      tags:
        - PlanNutricional
      summary: Eliminar PlanNutricional
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PlanNutricional eliminado
  /plan-nutricional/{id}/upload-image:
    post:
      tags:
        - PlanNutricional
      summary: Subir imagen a PlanNutricional
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida
  /progreso:
    get:
      tags:
        - Progreso
      summary: Listar todos los Progresos
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de Progresos
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/EntityBase'
    post:
      tags:
        - Progreso
      summary: Crear Progreso
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nombre
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '201':
          description: Progreso creado
  /progreso/{id}:
    get:
      tags:
        - Progreso
      summary: Obtener Progreso por ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Progreso encontrado
        '404':
          description: Progreso no encontrado
    patch:
      tags:
        - Progreso
      summary: Actualizar Progreso
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                nombre:
                  type: string
                descripcion:
                  type: string
      responses:
        '200':
          description: Progreso actualizado
    delete:
      tags:
        - Progreso
      summary: Eliminar Progreso
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Progreso eliminado
  /progreso/{id}/upload-image:
    post:
      tags:
        - Progreso
      summary: Subir imagen a Progreso
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: Imagen subida