# src/progreso/progreso.controller.ts
import {
  Controller,
  Get,
  Post,
  Body,
  Patch,
  Param,
  Delete,
  HttpCode,
  HttpStatus,
  Req,
  BadRequestException,
} from '@nestjs/common';
import { FastifyRequest } from 'fastify';
import { ApiTags, ApiOperation, ApiResponse, ApiBearerAuth, ApiParam, ApiBody, ApiConsumes } from '@nestjs/swagger';
import { ProgresoService } from './progreso.service';
import { CreateProgresoDto } from './dto/create-progreso.dto';
import { UpdateProgresoDto } from './dto/update-progreso.dto';
import { UploadService } from '../upload/upload.service';

@ApiTags('Progreso')
@ApiBearerAuth('JWT-auth')
@Controller('progreso')
export class ProgresoController {
  constructor(
    private readonly progresoService: ProgresoService,
    private readonly uploadService: UploadService,
  ) {}

  @Post()
  @HttpCode(HttpStatus.CREATED)
  @ApiOperation({ summary: 'Crear nuevo Progreso' })
  @ApiBody({ type: CreateProgresoDto })
  @ApiResponse({ status: 201, description: 'Progreso creado exitosamente' })
  @ApiResponse({ status: 400, description: 'Datos inválidos' })
  @ApiResponse({ status: 401, description: 'No autorizado' })
  async create(@Body() createProgresoDto: CreateProgresoDto) {
    const data = await this.progresoService.create(createProgresoDto);
    return {
      success: true,
      message: 'Progreso creado exitosamente',
      data,
    };
  }

  @Post(':id/upload-image')
  @ApiOperation({ summary: 'Subir imagen para Progreso' })
  @ApiConsumes('multipart/form-data')
  @ApiParam({ name: 'id', description: 'ID del Progreso' })
  @ApiBody({
    schema: {
      type: 'object',
      properties: {
        file: {
          type: 'string',
          format: 'binary',
        },
      },
    },
  })
  @ApiResponse({ status: 200, description: 'Imagen subida exitosamente' })
  @ApiResponse({ status: 404, description: 'Progreso no encontrado' })
  async uploadImage(
    @Param('id') id: string,
    @Req() request: FastifyRequest,
  ) {
    // Obtener archivo de Fastify
    const data = await request.file();

    if (!data) {
      throw new BadRequestException('No se proporcionó ningún archivo');
    }

    if (!data.mimetype.startsWith('image/')) {
      throw new BadRequestException('El archivo debe ser una imagen');
    }

    const buffer = await data.toBuffer();
    const file = {
      buffer,
      originalname: data.filename,
      mimetype: data.mimetype,
    } as Express.Multer.File;

    const uploadResult = await this.uploadService.uploadImage(file);
    const updated = await this.progresoService.update(id, {
      imagen: uploadResult.url,
      imagenThumbnail: uploadResult.thumbnailUrl,
    });
    return {
      success: true,
      message: 'Imagen subida y asociada exitosamente',
      data: { progreso: updated, upload: uploadResult },
    };
  }

  @Get()
  @ApiOperation({ summary: 'Listar todos los Progresos' })
  @ApiResponse({ status: 200, description: 'Lista de Progresos' })
  async findAll() {
    const data = await this.progresoService.findAll();
    return { success: true, data, total: data.length };
  }

  @Get(':id')
  @ApiOperation({ summary: 'Obtener Progreso por ID' })
  @ApiParam({ name: 'id', description: 'ID del Progreso' })
  @ApiResponse({ status: 200, description: 'Progreso encontrado' })
  @ApiResponse({ status: 404, description: 'Progreso no encontrado' })
  async findOne(@Param('id') id: string) {
    const data = await this.progresoService.findOne(id);
    return { success: true, data };
  }

  @Patch(':id')
  @ApiOperation({ summary: 'Actualizar Progreso' })
  @ApiParam({ name: 'id', description: 'ID del Progreso' })
  @ApiBody({ type: UpdateProgresoDto })
  @ApiResponse({ status: 200, description: 'Progreso actualizado exitosamente' })
  @ApiResponse({ status: 404, description: 'Progreso no encontrado' })
  async update(
    @Param('id') id: string, 
    @Body() updateProgresoDto: UpdateProgresoDto
  ) {
    const data = await this.progresoService.update(id, updateProgresoDto);
    return {
      success: true,
      message: 'Progreso actualizado exitosamente',
      data,
    };
  }

  @Delete(':id')
  @HttpCode(HttpStatus.OK)
  @ApiOperation({ summary: 'Eliminar Progreso' })
  @ApiParam({ name: 'id', description: 'ID del Progreso' })
  @ApiResponse({ status: 200, description: 'Progreso eliminado exitosamente' })
  @ApiResponse({ status: 404, description: 'Progreso no encontrado' })
  async remove(@Param('id') id: string) {
    const progreso = await this.progresoService.findOne(id);
    if (progreso.imagen) {
      const filename = progreso.imagen.split('/').pop();
      if (filename) {
      await this.uploadService.deleteImage(filename);
      }
    }
    await this.progresoService.remove(id);
    return { success: true, message: 'Progreso eliminado exitosamente' };
  }
}